Advanced async functions to offload db calls to cronjob
