+++ .env_template
@@ -3,3 +3,4 @@ DBUSER='<your_db_user>'
 PASSWORD='<your_db_password>'
 HOSTNAME='<your_db_host_ip_address>'
 DJANGO_SECRET='P@@@@@@@@@@@@@@$$$$$$$$$WWWW0000000000000oooooooorrrrrdddddd'
+DEBUG='True'
