+++ .gitignore
@@ -12,3 +12,4 @@ scrap/
 staticfiles/
 migrations/
 scratch.sql
+data/
