+++ static/modstyles.css
@@ -1,0 +1,70 @@
+
+.div-container {
+    display: grid;
+    grid-template-columns: repeat(1, 1fr);
+    gap: 10px;
+    margin: 10px;
+}
+
+.odd { 
+    background-color: #d9edf7; 
+}
+
+.generated_for_mobile { 
+    margin-bottom: 30px 
+}
+
+
+/*
+Block for showing meetings table 
+on mobile devices specifically.
+*/
+
+@media only screen and (max-width: 760px) 
+            and (max-device-width: 1024px)  {
+
+
+table, table-striped, table-bordered, table-hover, 
+thead, tbody, th, td, tr { 
+    display: block; 
+}
+
+/* Hide table headers (but not display: none;, for accessibility) */
+thead tr { 
+    position: absolute;
+    top: -9999px;
+    left: -9999px;
+}
+
+tr { border: 1px solid #959697; }
+
+td { 
+    /* Behave  like a "row" */
+    border: none;
+    border-bottom: 1px solid #eee; 
+    position: relative;
+    padding-left: 50%; 
+}
+
+td:before { 
+    /* Now like a table header */
+    position: absolute;
+    /* Top/left values mimic padding */
+    top: 6px;
+    right: 6px;
+    width: 45%; 
+    padding-right: 10px; 
+    white-space: nowrap;
+}
+
+/*
+Label the data
+*/
+td:nth-of-type(1):before { content: ""; }
+td:nth-of-type(2):before { content: "Meeting Name"; }
+td:nth-of-type(3):before { content: "Virtual"; }
+td:nth-of-type(4):before { content: ""; }
+td:nth-of-type(5):before { content: "Start Time"; }
+td:nth-of-type(6):before { content: "Duration"; }
+td:nth-of-type(7):before { content: "Formats"; }
+}
